pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "FastAPI Template - Production-ready FastAPI template with PostgreSQL, Redis, and modern Python tooling"
requires-python = ">=3.13"
dependencies = [
    "greenlet>=3.2.4",
    "psycopg[binary]>=3.2.10",
    "sqlmodel>=0.0.27",
]

[dependency-groups]
api = [
    "fastapi[standard]>=0.119.0",
    "httpx>=0.28.1",
    "bcrypt>=5.0.0",
    "uvicorn>=0.37.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "authx-extra>=1.2.0",
    "authx>=1.4.3",
    "structlog>=25.4.0",
]

# migration-only
migrate = [
    "alembic>=1.17.0",
]

dev = [
    "coverage>=7.10.7",
    "debugpy>=1.8.17",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "types-passlib>=1.7.7.20250602",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.13"
disable_error_code = ["arg-type", "assignment", "import-untyped", "call-overload"]
warn_return_any = false
warn_no_return = false
check_untyped_defs = true
exclude = ["venv", ".venv", "tests", "alembic", "data"]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["venv", ".venv", "alembic", "tests", "data"]

[tool.ruff.lint]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$|kwargs|args"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
    "UP040", # Allow TypeAlias annotation (required for FastAPI dependencies)
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
